requirements.txt
absl-py==1.1.0
altair==4.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
beniget==0.4.1
bleach==5.0.0
boto3==1.24.9
botocore==1.27.9
category-encoders==2.5.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.4
cryptography==37.0.2
cycler==0.11.0
Cython==0.29.30
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.3
gast==0.5.3
great-expectations==0.15.9
htmlmin==0.1.12
idna==3.3
ImageHash==4.2.1
importlib-metadata==4.11.4
iniconfig==1.1.1
ipykernel==6.13.1
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.3
llvmlite==0.38.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
missingno==0.5.1
mistune==0.8.4
multimethod==1.8
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
notebook==6.4.12
numba==0.55.2
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
pandas-profiling==3.2.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.1.1
pluggy==1.0.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==4.21.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pybind11==2.9.2
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyparsing==2.4.7
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
pythran==0.11.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.1.0
requests==2.28.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
scikit-learn==1.1.1
scipy==1.8.1
scour==0.38.2
seaborn==0.11.2
semantic-version==2.10.0
Send2Trash==1.8.0
shap==0.41.0
six==1.16.0
slicer==0.0.7
soupsieve==2.3.2.post1
stack-data==0.2.0
statsmodels==0.13.2
tangled-up-in-unicode==0.2.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
typing-extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
verifyml==0.0.6
visions==0.7.4
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.8.0
